feat: 알림 컴포넌트 분리 및 라우팅시 재요청 및 UI변경
